refactor(svar2): split VcfChunkReader into RecordSource + ChunkAssembler
VcfChunkReader fused htslib iteration, atom decomposition, the reorder heap,
presence packing, and field staging; only the first is VCF-specific. Extract the
source-agnostic spine into ChunkAssembler behind a RecordSource trait yielding an
owned RawRecord, leaving VcfRecordSource as the htslib adapter. No behavior
change -- SVAR2 output is byte-identical.
