+++ src/lib.rs
@@ -17,6 +17,8 @@ pub fn bits_get_bit(bytes: &[u8], i: usize) -> bool {
 }
 #[cfg(feature = "conversion")]
 pub mod budget;
+#[cfg(feature = "conversion")]
+pub mod chunk_assembler;
 pub mod cost_model;
 pub mod dense;
 #[cfg(feature = "conversion")]
@@ -55,6 +57,8 @@ pub mod py_query_batch;
 pub mod py_query_decode;
 pub mod py_query_ranges;
 pub mod query;
+#[cfg(feature = "conversion")]
+pub mod record_source;
 pub mod rvk;
 pub mod search;
 pub mod spine;
